Add . childPenguinTest

+++ src/main/java/ChildPenguin.java
@@ -1,9 +1,9 @@
 public class ChildPenguin implements Penguin {
-    final int charge = 170;
+    final float charge = 170;
     final int percentage = 3;
 
     @Override
-    public int calculate() {
-        return (int) (charge * (percentage / 100.0f)) + charge;
+    public float calculate() {
+        return (charge * (percentage / 100.0f)) + charge;
     }
 }
