+++ src/main/java/FemalePenguin.java
@@ -1,9 +1,11 @@
+import java.text.DecimalFormat;
+
 public class FemalePenguin implements Penguin {
-    final int charge = 145;
+    final float charge = 145;
     final int percentage = 7;
 
     @Override
-    public int calculate() {
-        return (int) (charge * (percentage / 100.0f)) + charge;
+    public float calculate() {
+        return ((charge * (percentage / 100.0f)) + charge);
     }
 }
